Drop InputRedirectionCursor
The goal of this change is to drop scary cursor synchronization code in
the InputRedirectionCursor type by dropping it.

With the proposed design, the PointerInputRedirection helper directly
manages the position of the cursor.

If the cursor (and the pointer) needs to be moved, one should use the warp
function in the PointerInputRedirection.

Cursors::removeCursor() was touched to avoid the mouse cursor destructor
potentially accessing released memory.