Hook up new output props allowDdcCi and detectedDdcCi
Requires a new plasma-wayland-protocol revision with updates to
output-device-v2, output-management-v2 and external-brightness-v1.
This introduces the "ddc_ci" capability and a "ddc_ci_allowed" flag
for outputs, as well as "set_uses_ddc_ci" for brightness devices.

KWin will store "allowDdcCi" and "detectedDdcCi" in its config.

If "detectedDdcCi" is set to true, based on the brightness device's
usesDdcCi flag, then it will be remembered and exposed as "ddc_ci"
capability. This in turn can be used by a settings UI to determine
whether the "allowed" flag should be configurable.

If "allowDdcCi" is set to false, contradicting the default value,
then a brightness device with usesDdcCi flag will not be assigned
to an otherwise suitable output. KWin will use software brightness
transformations instead.