backends/drm: Update dpms mode when dpms off is cancelled before animation completes
When the turn off timer starts, the dpms mode is changed to AboutToTurnOff.
If the dpms mode is changed back to `On` before the animation completes,
we need to update the dpms mode from AboutToTurnOff to On.

(cherry picked from commit a4b351d5b88d53ca7326da2a34ca4b65afe57eb7)