Make outline print scene graph error
This is to prevent the outline potentially crashing kwin when a graphics
reset occurs.

If the QSGRenderLoop fails to create an OpenGL context (can happen during
graphics reset), it will check whether the QQuickWindow::sceneGraphError
signal is connected. If it is, the QSGRenderLoop will emit the signal,
clean up some things and continue execution. If it is not, the
QSGRenderLoop will crash the app.


(cherry picked from commit 4a8020a3f502f6ee9593c51780f728fc6ad44dda)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>