wayland: implement wp-fifo-v1
This allows apps to get proper fifo behavior and queue up multiple frames for rendering,
as well as not being blocked on eglSwapBuffers / vkQueuePresent with swap interval 1 /
FIFO presentation mode.

How it works is that on each commit, the client can
- set a barrier, which does nothing on its own
- wait on a (potentially) previously set barrier

The barrier on the surface is removed every time the surface gets presented, which unlocks
the next transaction/commit to be applied. The Mesa implementation sets the barrier and wait
on every commit, ensuring that every frame from the application is presented for at least one
refresh cycle.