wayland: Rework transaction fences
The goal of this redesign is to address fundamental architectural flaws
in the transactions that can result in unexpected application of surface
state during client tear down.

For example, if a transaction contains several surfaces (the main surface
and one or two subsurfaces) and the transaction is blocked on one of
the surface, the user may observe different outcomes after closing the
windows. For example, sometimes everything might be alright, sometimes
you may see glitches. Another problem with it is that it makes harder to
reason about the compositor's behavior. It is easier to reason in terms
such as "a transaction will be applied when all the dependencies are
satisfied" and assume that nothing is going to change when the client
connection goes down.

Another reason for this redesign is to pave the way for transaction
poisoning or dropping certain transaction states because the surface role
got destroyed.

Transaction fences are associated with particular transaction entries.
The reason for it is to avoid blocking the entire transaction because of
an entry whose state is not going to be applied.

If the whole client connection goes down, then the transaction DAG will
be cleaned up and that's all, no surface state is going to be touched.

If only one surface is destroyed and the act of destroying the surface
unblocks the transaction, then the transaction will be applied. It may
produce visual glitches, but we have no other better alternative at the
moment.

Transaction locking and unlocking is dropping until there is a user again.