wayland/colormanagement: ignore frog surfaces for the surface_exists error
SDL uses frog color management for the feedback part only, but when a newer version
of Mesa is combined with an older version of SDL, it gets this protocol error.

This prevents that issue by just removing the error case. Longer term, this gets
solved more nicely by just dropping the frog protocol.