wayland: Drop ClientConnection::disconnected
The main goal of this change is to simplify the client connection by
reducing its surface api.

The aboutToBeDestroyed signal is emitted when the client connection is
about to be destroyed, it is emitted before any wl_resource owned by
the client is touched.

The destroyed signal is emitted when the deed is done, i.e. the client
connection and all its resources have been destroyed.