utils: Add operator bool to FileDescriptor
Its intent is to indicate whether the file descriptor is valid. This can
be used to clean up if statements, for example

    if (FileDescriptor fd = somethingSomethingFile(); fd.isValid()) {
        // use the fd
    }

can be changed into

    if (FileDescriptor fd = somethingSomethingFile()) {
        // use the fd
    }

which is more convenient.

The only downside is that if you look at the file descriptor only just as
a number, then it doesn't make sense. But, one could also argue that this
use case is invalid or unrealistic because file descriptor 0 is valid.
There is no good reason to special case file descriptor 0 like that, such
code is very likely buggy.

The `explicit` keyword is used to prevent the compiler from implicitly
converting the boolean in an integer, so if you still need to pass a file
descriptor number, you must use the FileDescriptor::get() function.