wayland: Clean up how the effective buffer damage is computed
This change re-arranges the code so surfaceRect and bufferRect are
computed just once. It also changes things a bit so watching the
damaged signal is not an absolute requirement. For example, one could
inspect the damage() property after the committed() signal is emitted.
If it's empty, there is no damage; if it's not empty, then the surface
contents has been damaged. Currently, you can't write code like this.