wayland: Sanitize surface damage when surface is unmapped on commit
The damage cannot be reset when processing wl_surface_attach() because
we don't know for sure whether there is going to be future
wl_surface_damage() or wl_surface_damage_buffer() request, or whether
the client is actually going to attach a valid buffer after all, e.g.

   wl_surface_damage(surface, region)
   wl_surface_attach(surface, nil)
   wl_surface_attach(surface, buffer)

or

   wl_surface_damage(surface, region)
   wl_surface_attach(surface, nil)
   wl_surface_damage(surface, region2)

Resetting the damage at the surface commit time is a more reasonable
thing to do.