wayland: Interlock surface state with surface extension state
The rules for the pending surface extension state are relaxed currently.
Consider, the following surface state queue

   A <- B <- C <- D <- E

If an extension is created when C is the pending state, the extension
will process commits A and B. With the proposed changes, only states
starting from C will be processed by the extension.