wayland: Rework surface attached state
The rules for the pending surface extension state are relaxed currently.
Consider, the following surface state queue

   A <- B <- C <- D <- E

If an extension is created when C is the pending state, the extension
will process commits A and B. With the proposed changes, only states
starting from C will be processed by the extension.

In order to make extension state <-> surface state association robust,
this change makes the SurfaceExtension store its state in the SurfaceState.
This guarantees that state tracking doesn't break when the pending state
is stashed or if a synchronized subsurface is committed multiple times
before the parent surface.