wayland/colormanagement: default max luminance to the max. transfer function luminance
The protocol says that this should be done:

"If mastering luminances are not explicitly set, the target color volume is assumed to
have the same min and max luminances as the primary color volume"

and for PQ the resulting tone mapping looks a bit less bad than clipping. For SDR content
and scRGB it doesn't make a difference.