Reroute all internal window geometry updates through InternalWindow
Currently, the QWindow is fully in charge of window geometry updates,
which creates a series of issues for kwin revolving around breaking
update cycles.

 * If an internal window wants to update its geometry, it will call
   QWindow::setGeometry(). kwin, in its turn, will update the geometries
   in the InternalWindow and call QWindow::setGeometry() again, and
   hopefully that's where the geometry update ends

 * If kwin wants to update the geometry of an internal window, it will
   call QWindow::setGeometry(). When kwin sees that the geometry of
   the QWindow changes, it'll call QWindow::setGeometry(), and fingers
   crossed that that's where the update sequence ends

The geometry update process is very clumsy and it is so because kwin
has no idea why the geometry of a QWindow changes. Is it initiated by
kwin or the client?

This change re-routes the geometry requests through InternalWindow. If
the client wants to update the geometry, the InternalWindow must process
that request.

 * When the InternalWindow has finished processing a geometry request,
   it will call QWindowSystemInterface::handleGeometryChange() to notify
   the window about new geometry
 * The internal window then repaints
 * When a frame is rendered, the current geometry is updated