Only forward tablet events to the focused window when the tool is down
Currently we forward events to the top-level surface at the event position,
which results in all sorts of unwanted behavior. One example is when you start
drawing with a pen it can easily activate other windows or applets, like the
hover in Task Manager.

Instead, we should create an "implicit grab" when the tool tip is down. This
acts just like a pointer does, and "locks" the tablet input into the current
window.