screenedge: Hardcode the cornerOffset value
At the moment we were using fonts to calculate it which would slow us
down at startup for little pay-off.
We were using visual metrics for a touch scaling which is wrong so let's
just hardcode a value.
On my system (running default fonts) gridUnit was 9 and thus was
resulting on 36, rounded up to 40.