wayland: Apply surface extension state before emitting changed signals
This fixes missing window close animation in chromium. The
XdgToplevelInterface needs to process the new state as soon as possible
so it can notify the Window that the xdg_toplevel surface has been
unmapped and the WindowItem needs to be frozen.

If the WindowItem is not frozen in time, then surfaces in the SurfaceItem
subtree may update their state to reflect unmapped state, i.e. they'll
become invisible.