Simplify xwayland surface association
When kwin starts managing or tracking an X11 window, there should not be
any wl_surface associated with the X11 window.

While on this, also slightly rewire the code so the associate() function
is not called indirectly by setSurface(). It allows us to find all relevant
places where the wl_surface is associated by using code navigation tools
such as "find all references".