wayland: implement the single pixel buffer protocol
The protocol does exactly what it sounds like, it allows the client to create
buffers that are only a single pixel large, in a simpler and more efficient way
than creating a 1x1 shm buffer.

To keep things simpler for now, this makes the renderer just upload a 1x1 texture
with the desired color. This can later be improved on by special casing single
pixel buffers in the renderer.