core/output: make the uuid properly persist
Before, it was generated with information from the edid and the connector name,
which meant that it could change depending on how you connected the display.

The new ID is saved and loaded in the output configuration system, which means
that other settings using the ID will now be properly persisted whenever the
output can be uniquely identified at all.