outputconfigurationstore: implement better screen mirroring
Instead of KScreen trying to move and scale outputs in a way that kind of looks
like mirroring, this makes KWin aware of which outputs replicate which, and makes
it responsible for implementing mirroring in whatever way is best.

This commit implements simliar workarounds like KScreen had, with the only real
difference being that it avoids modifying the source display mode and position,
and takes screen rotation into account.
A better implementation can replace that bit later on, while making use of the
same information stored in the configuration.

BUG: 502630
CCBUG: 481222