backends/drm: always signal the available HDR headroom on laptop displays
Currently on SDR laptop displays, we send clients image descriptions with values like
- reference luminance 200
- transfer function max. luminance 200
- max. mastering luminance 200

When any content using enough HDR headroom is shown, we change that to something like
- reference luminance 200
- transfer function max. luminance 300
- max. mastering luminance 300

However, if the client decides whether or not to show HDR content depending on display
capabilities (if max. mastering luminance > reference luminance), then this situation
will never happen. To fix that, we have to advertise the possible max. luminance to the
client before we can actually show it, resulting in values like
- reference luminance 200
- transfer function max. luminance 200
- max. mastering luminance 300