Track window placed state
This fixes chromium's window position not being correctly restored.

Chromium destroys the xdg toplevel session after receiving the restored
event. This breaks XdgToplevelWindow::isPlaceable(), it may return
true and Workspace::addWaylandWindow() will overwrite the restored
position.

If the window position is restored, the Workspace should avoid placing
the window even in case the toplevel session has been destroyed.

With the proposed changes, the Window type gains a new state to
indicate whether the window has been placed. If the window position is
restored from the session, that property will be set and the Workspace
will avoid placing the window.