Fix mouse cursor not changing when zoomed in
This commit fixes an issue where the zoom effect never gets
notified of cursor texture changes when the current cursor
source is a SurfaceCursorSource.

The bug is caused by the fact KWin will only allow
SurfaceCursorSource to commit surface updates while the
cursor is visible. Zoom effect hides the cursor so you
don't see two of them, but it still needs the surface
updates.

This commit removes the visibility check.

BUG: 492173
Signed-off-by: Ritchie Frodomar <alkalinethunder@gmail.com>