Add PlacementCommand
The place methods in the Placement type move the window. However, with
the XdgToplevelWindow, there are cases where we just would like to check
where the window is going to land without the move() function getting
called.

The new PlacementCommand helper provides a more declarative approach to
the window placement.