outputconfigurationstore: disable autorotate when screen is off
If device does not have screen on, there is no need to process the
screen rotation events coming from sensors. For mobile devices this
results in the quite a lot of CPU usage for applications when screen is
off.

This most likely is not case when used with laptops and tablets but for
mobile devices this can be very significant when device is in pocket.
Each leg movement may end up becoming a rotation event when walking or
climbing stairs for example. On mobile devices sensors are very
sensitive to such an small changes in orientation.

This is tested on Oneplus 6 with postmarketOS edge.