Drop X11-specific hack in shouldKeepTransientAbove()
This hack is needed to prevent kickoff configuration dialog being a proper
transient child of the panel (it's a group transient), otherwise the dialog
will be unreasonably kept above all other windows.

On Wayland, a dock window must be a layer shell window, and a layer shell
window cannot be a parent of a dialog. And even if such a case were allowed,
we'd need to fix plasma so it doesn't set up a nonsensical window hierarchy.
On X11 though, it's hard to fix.