autotests/x11keyread: fix race condition
xcb_flush only ensures that the request is written to the socket, it does not
mean that the X server has actually proccessed the request. Sometimes, that
means the test fails, because the key events get sent and processed before
the x11 event recorder is set up - making the test randomly fail.

This commit replaces it with Xcb::sync, which blocks until the X server has
processed the event, and which should make sure the test works reliably.