Create implicit grab when tablet tool button is down
This is a follow-up to comments about cf676b86db59889a2af3122746b699d6cd055cd9
where it's suggested that tool buttons should also trigger an implicit
grab. This makes sense, as for example Krita uses the first tool button
to pan the canvas by default. Just like with drawing, it's easy to "pan"
outside of the window and break the panning operation.