draft: improve handling of dual mode "tiled" displays
Some displays, like the Apple Studio Display, can operate in both "normal" and
tiled mode. They expose multiple tiles, but each tile also supports the native
resolution of the full display.
Previously, the display would be exposed to the user as two separate connectors;
with this commit, the tiles aside from the one at position 0, 0 are filtered out
and the native mode is preferred.

TODO:
- keep the filtered-out tiles disabled
- adjust output config store to take tile info into account for the native mode
- filter out the outputs for output management too