effect: Allow AnimationEffect running while the screen is locked
The restriction was introduced in 2a9971fa080876c21332049a0b249c43cc9764c9
to prevent the window aperture effect showing windows in screen corners
on the lockscreen.

The restriction is very broad. It breaks the phone app in plasma mobile.

Since the visuals of the window aperture had been changed so the windows
remain transparent in the screen corners, this restriction can be
removed. API-wise, if we need such a behavior again, it should be
implemented differently, for example by providing a metadata key to tell
kwin that it is fine to run the effect while the screen is locked or
making the effect cancel its animations after locking the screen.