plugins/glide: don't grab window add/close roles
This fixes the inconsistency that FallApart can be enabled
with Fade and Scale, and the closing animation would be
FallApart. But it doesn't work with Glide.

Neither Fade or Scale grab the Window{Added/Closed}GrabRole.
It makes sense as they have animations both for opening and
closing. So if the user enables one of them, _and_ a opening/
closing-only effect, he would expect the latter to be preferred.

This commit makes the three effects in a group consistent.