backends/drm: remove GraphicsBuffers from the cache when they're deleted
While DrmFramebufferData cleans up after itself, if it's still cached in DrmPlane
when the framebuffer gets deleted, it outlives the buffer. If afterwards a new
GraphicsBuffer is allocated with the same pointer address, we can end up re-using
the old DrmFramebufferData with that new different GraphicsBuffer, showing an old
frame instead of the actual buffer contents.