plugins/nightlight: Fix conversion of julian date to QDateTime
The timezone shouldn't matter when night light is synchronized to the
Sun. Currently, it's not the case though.

For example, if you set your location to Beijing but still use the
Europe/Kyiv timezone, the night light will be active even though it's
daylight in China, e.g. at 10:00AM in UTC+3.

This change addresses the issue by converting julian dates to QDateTime
as is instead of making assumptions that julian days will match both
in the final and in the input julian dates.

It fixes night light unexpectedly getting activated after flying to
another country and not adjusting your timezone.