Reset FIFO status on each surface state
A FIFO barrier is valid up until the next commit, if a new FIFO barrier
is needed the caller should create a new barrier. It is not a state that
should persist on the surface.

This fixes Qt Vulkan clients after a hide/show, as they block
indefinitely waiting for the surface to be configured as kwin
does not process the empty commit waiting on a barrier.