plugins/nightlight: Add SunTransit helper
The SunTransit helper provides a way to compute the timings of sunset,
sunrise, etc. It provides slightly more accurate timings, at least for
my location. All the timing calculation code is abstracted in a way that
should allow moving the code in a shared library to avoid duplication
between kwin and night light kcm.

The hour angle values are capped to make sure that the SunTransit's
behavior is reasonable. With extreme hour angles, the timings are less
predictable, for example

    SunTransit(dateTime.addDays(1), latitude, location).dateTime(CivilDawn)

can return a QDateTime that preceeds the dateTime.

More math details can be found at https://gml.noaa.gov/grad/solcalc/calcdetails.html.