wayland: Ignore requests from former data offers during dnd
If a wl_data_offer_accept() with a nil mime type arrives from a former
data offer, it can interfere with the current data offer.

This change makes the data device track data offers. If the data device
leaves a surface, the corresponding data offer will be destroyed. However,
if it is after a drop, then the data offer will be kept around so the
target client can retrieve data.

BUG: 504089