cppcheck: better KWIN_SINGLETON definition
Rather than telling cppcheck it should treat it as a empty define, teach
it what it actually resolves to.

Given this is a code generating macro that actually influences other
things, we should do it properly.

This is macro creates the constructor, and without cppcheck knowing
that, we have a class without constructors, so everything is
unitialized.