plugins/qpa: Send Move and Resize events for invisible windows
This is to ensure that QWindow size is properly synchronized with QMenu size.

If the QWindow screen changes (it can happen if the menu is going to appear
on a screen other than QGuiApplication::primaryScreen()), QWindow will call
QGuiApplication::processGeometryChangeEvent(). It may generate a resize event
and change the size of the QMenu.

If there's a follow up setGeometry() for the QWindow, a resize event must
be generated otherwise the QMenu will get stuck with a wrong size. QtXCB
operates in a similar fashion, it'll call handleGeometryChange() even though
the window is hidden.

BUG: 498111