x11: Skip updating input shape when bounding shape changes
This should be unnecessary.

It also fixes the X11Window emitting the shapeChanged() signal when the
input shape changes. The shapeChanged() signal should be emitted only
when the bounding shape changes, which is relevant for rendering.

detectShape() can be potentially merged with updateBoundingShape(), but
it's somewhat tricky to do because the manage() function should be updated
too and I'm not sure what the proper initialization sequence should look
like there.