tiles: Fix CustomTile::remove() collapsing tile tree too much
When a tile is removed and its sibling is empty, the parent tile will
be converted in a leaf tile.

The problem is that CustomTile::remove() can attempt to convert the
root tile in a leaf tile, but we expect the root tile to contain at
least one child at all times.

This change adds a relevant guard. Note that CustomTile::remove() can
be still used to purge all children of the RootTile. This is mainly
used by the tiles editor. Window::quickTileGeometry() was also adjusted
so it doesn't report a misleading geometry if the window can't be
custom tiled.

BUG: 504367


(cherry picked from commit 9b0fd06e14e30c8fa39271fe6b1c0faaee381588)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>