x11: Remove bounding shape fallback for noborder state
Currently, the bounding shape is taken into account when determining
whether a window decoration can be drawn on application.

kwin does so because of applications such as xeyes. However, there is
a correct way to communicate to the window manager that no window
decoration should be drawn using motif hints. For example, both GTK and
SDL set it.

This change removes the bounding shape fallback with a couple of things
in mind: first, it helps us with converting kwin from a reparenting
window manager to a non-reparenting window manager (with this, kwin will
not need to forward some client window properties, etc); and this makes
X11Window code simpler and more maintainable.

As far as I know, no spec mandates that the window manager mustn't
draw a frame around a shaped window.

There is a risk of breaking some applications, but I expect that most of
the affected applications will be such as xeyes. In any case, we can
go with this change, and if there are other major usecases that we have
not accounted for, it can be reverted.

It is also worth noting that Mutter doesn't have such a fallback either.
So if an application wants to have no border on Linux, it already must
use motif hints.


(cherry picked from commit 45eff5be3b0873ac4722ca2e7d0d1034321726d1)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>