stroke: StrokeGesture and StrokeGestures are now our C++ wrappers
The original EasyStroke C code, and its stroke_t struct,
remains hidden away in the implementation file.

StrokeGesture, with a triggered() signal, replaces Stroke as the
primary C++ wrapper class. Stroke is now only a typedef for an
opaque smart pointer type of the C struct, and is only necessary
in the API for comparison with StrokeGesture objects.