kcms/xwayland: Rephrase things to be clearer
Right now this KCM has a wall of text on top, and no guidance for the
implications of using the individual options. This isn't ideal.

Instead, let's reduce the text on top to only the minimum needed to
explain what this KCM is for, and then instead offer inline
explanations for the implications of each individual setting.

FEATURE: 504395
FIXED-IN: 6.5.0