Remove shading leftovers
Shading is not implemented on Wayland because proper asynchronous window
geometry updates are very very hard to achieve. With shading, the code
will be beyond our capabilities what we can maintain or extend. At quick
glance, shading is just about hiding the window contents, but in reality,
it interacts with other features in lots of small ways. Those features
have to accommodate for shading, some Wayland code paths used to have such
code but they resulted in window geometry bugs, for example windows being
too small, the only feasible option at the time was to remove that code
because otherwise the code complexity was too unbearable to fix the bugs.

That leaves us with a consistency problem though: X11 windows sort of
support shading (because shading is still supported by the Xorg session;
although to be fair it's quite buggy), but Wayland windows do not.

After kwin_x11 and kwin_wayland split, the corresponding system settings
had been removed in kwin_wayland. One of the reasons for the split was
exactly like this where we cannot deliver a feature on Wayland but also
don't want to break even more things for X users. This change follows up
on the split by removing the remaining shading bits in kwin_wayland, which
are not exposed in system settings.