Unify name for the "window menu" and its button
This commit unifies the user-visible name for what I've called the
"window menu", and its associated button. This is the button that is
by default in the top-left of a window with the application's icon,
and that shows the menu with Maximize, Minimize, Close etc. when clicked.

At least at some point in the past, this was simply called the "Menu
button" and that text was still in use in some places. More recently the
name for the button (shown in the tooltip and in the Titlebar Buttons KCM)
was changed to "More actions for this window". The keyboard shortcuts KCM
called it "Window Operations Menu". The window behavior KCM called it the
"Actions Menu".

Now let's unify on "window menu".

BUG: 499700
BUG: 504807
FIXED-IN: 6.5.0