scene/item: restrict frame callbacks based on the output rather than geometry
Sending frame callbacks based on whether or not the bounding rect intersects with the
rendered output has two big problems:
- when a window is entirely off-screen, it never gets frame callbacks. This is particularly
  problematic when the application waits for frame callbacks to render the next frame, but
  KWin waits with moving them to an output until the application commits the surface
- when a window is on two outputs at once (like with cloning), then it gets twice the
  amount of frame callbacks - which means the application ends up rendering twice as
  fast, which doesn't make any sense

This commit replaces that approach with just checking if the output is the same as the one
Workspace picks for the center of the window. If the window is entirely off-screen, that
will still result in an output getting chosen, and it also can't be on two outputs at once.

BUG: 479694
BUG: 498628
BUG: 505060