compositor: rearrange compositing to test all layers atomically, and only render afterwards
This reduces the amount of atomic tests per presentation and makes it more feasible to add more
layers than just primary and cursor.
This commit also unifies code for primary and cursor layer painting, and opens up the possibility
of doing damage tracking on the cursor plane (currently disabled though because of bugs).