plugins/blur: Make background flash less strong
With the current modulation curve, the background can be more opaque
than the window. Since the contrast effect raises the brightness a bit,
it can result in a flash if the window has dark contents.

This patch changes the modulation curve to `alpha ** 2`, so the
background contents is less strong. There is still a bit of flashing
but it's barely noticeable now. The only downside of this change is
that windows can lose their "background" faster when they are animated,
but in my testing, it is not a really noticeable issue.

BUG: 505440