backends/drm: refactor the implicit modifier fallback
We do want to fall back to implicit modifiers if explicit ones don't work, but how that
was implemented wasn't ideal - it required storing and modifying an additional formats
+ modifier map, and caused a lot of unnecessary re-allocations and additional tests
when an output configuration isn't possible.

This commit changes it to fall back to implicit modifiers on all planes at once, and for
all formats at once, which is faster and should be just as effective.