plugins/highlightwindow: Highlight parent windows together with modal dialogs
Modal dialogs are a bit special. If there is a modal dialog, it will be
activated when the user attempts to activate the parent window.

As of now, the task manager asks kwin to highlight parent windows, the
modal dialogs are not included. It can be unexpected that when you click
the parent entry, the modal dialog will actually be activated.

This change makes the highlight window also include modal dialogs so
they are not activated out of the blue.

CCBUG: 322087