backends/drm: tie output layers more closely with their drm planes
This is needed to more properly deal with overlay planes, where we may have only
one for all outputs, or where we can switch them between outputs at will.
This also makes the Compositor deal with cursor planes on some but not all crtcs.
That's not something we've really seen in practice so far, but is theoretically
allowed.