Prepare for the ability to ignore key repeat events in global shortcuts
This commit sets up some of the necessary work on kwin's side to allow
for global shortcuts registed with kglobalaccel to choose to ignore
the user continuously holding the shortcut.

When the rest of the work is done, this will allow effects like the
Invert effect to stop continuously flashing the screen when the "Toggle
color Inversion" key is held.

The Invert effect now calls QAction::setAutoRepeat(false) on its two
toggle actions, but that currently has no effect. This needs to be addressed
in kglobalaccel.

Signed-off-by: Ritchie Frodomar <alkalinethunder@gmail.com>