scripting/workspace_wrapper: (505688) make workspace.desktopGridHeight writable
https://bugs.kde.org/show_bug.cgi?id=505688

This MR makes the VirtualDesktopManager grid *height* writable from scripts.

Since the grid size is computed in VirtualDesktopManager::updateLayout from the height and virtual desktops count, I thought it makes sense to only provide a setter for the height.

We can already change the virtual desktops count from scripts, this allows changing the height as well.

Here is a link to a usecase for this feature: https://github.com/maurges/dynamic_workspaces/issues/4

Thanks,
Dorian.