compositor: make use of overlay planes
Whenever a quickly updating Item is unobstructed and doesn't have any effects on or above
it, we attempt to put it on an overlay layer. This allows skipping compositing in many
situations, which should noticeably improve efficiency and performance.

If the compositor can't put all the quickly updating Items in the scene on overlays for any
reason, it falls back to compositing on the primary plane. More complicated heuristics for
putting some but not all Items on overlays could be added later, but for now the focus is
purely on efficiency and keeping the code simple.

As it's anticipated that this feature will cause problems on some drivers or hardware, the
environment variable KWIN_DISABLE_OVERLAYS can be used to turn off overlay plane usage.