compositor: make use of overlay planes
Whenever a quickly updating Item is unobstructed and doesn't have any effects
on or above it, we attempt to put it on an overlay layer. This allows skipping
compositing in many situations, which should noticeably improve efficiency and
performance.

If the compositor can't put all the quickly updating Items in the scene on
overlays for any reason, it falls back to compositing on the primary plane.
More complicated heuristics for putting some but not all Items on overlays
could be added later, but for now the focus is purely on efficiency and keeping
the code simple.

As it's anticipated that this feature will cause problems on some drivers or
hardware, it's only enabled by default in development versions. The environment
variable KWIN_USE_OVERLAYS can be used to still enable overlay plane usage in a
release version for easier testing, or disable it in a development version to
temporarily work around issues.