Fix Window::mousePressCommandConsumesEvent() for "activate" action
Amends 87bb2447e5328a5e6f4d066fcfea5ff38a02139d.

According to the boolean algebra rules, if the `a || b` expression is
negated, the result will be `!a && !b`.

In the prior version, the MouseActivate option used to correspond to
`isActive() || !rules()->checkAcceptFocus(acceptsFocus())` so following
the boolean algebra rules, the `||` operator should be changed to `&&`.

BUG: 506007