core,opengl: allow the tone mapper to reduce reference luminance more
I assumed that it was fine before and wanted to be more conservative, but with high
brightness videos, the 50% additional range crushes highlights way too much.
This is especially visible in videos with snow, where most of a frame can be between
200-500% reference luminance, which were before compressed to like 10% of the total
brightness range.
This commit adjusts the tone mapper to reduce the reference luminance by half, instead
of the previous 2/3. This should be good enough for most videos, without making games
too dark.