compositor: clamp artificial hdr headroom to the currently max possible one
When you increase brightness, the hdr headroom is immediately reduced, but the
color description was only adjusted with an animation, which ended up clipping
highlights until the animation is complete. This immediately adjusts the color
description instead.
If for example the previous brightness level was 1/3, and the artificial hdr
headroom 3.0, but now you increased brightness to 100%, then the artificial hdr
headroom was still 3.0 before, and only slowly got animated towards 1.0. With
this commit, it's immediately adjusted to be 1.0 instead.