effect: Allow AnimationEffect::redirect() change all termination flags
Currently, the redirect() function forbids changing the TerminateAtTarget
flag. But it can be useful if a set() animation needs to be canceled in the
middle of progress.

This change allows effects to change all termination flags in redirect().
Effectively, it allows changing the animation type at runtime, i.e.
switch back and forth between animate() and set() types.