plugins/highlightwindow: Prevent leaking revert animations
The revert animation leaks an animate() animation. If the highlight
window schedules a new set() animation, it will make the window opacity
value change in an unexpected way.

This change reworks the highlight window effect so it reuses the same
set() animation when transitioning from the normal state to the
ghost or highlighted state and vice versa.